renamed "src/parsers/objects/schemas" & "tests/test_parsers/test_objects/test_schemas" to "src/schemas" & "tests/tests_schemas"
