+++ src/schemas/array.py
